Create composer.js
Created new file and added logic to show a button at the bottom of the page when the user starts to scroll and make the nav button disappear. When the button at the bottom is clicked the page scrolls to the top and the form to add a new tweet slides down.